+++ src/lib/waku_relay/index.ts
@@ -121,7 +121,7 @@ export class WakuRelay extends Gossipsub {
    */
   public async send(message: WakuMessage): Promise<void> {
     const msg = message.encode();
-    await super.publish(this.pubSubTopic, Buffer.from(msg));
+    await super.publish(this.pubSubTopic, msg);
   }
 
   /**