+++ src/test_utils/constants.ts
@@ -5,5 +5,22 @@
  * @module
  */
 
-export const NOISE_KEY_1 = Buffer.alloc(32, 1);
-export const NOISE_KEY_2 = Buffer.alloc(32, 2);
+export const NOISE_KEY_1 = new Uint8Array(
+  ((): number[] => {
+    const b = [];
+    for (let i = 0; i < 32; i++) {
+      b.push(1);
+    }
+    return b;
+  })()
+);
+
+export const NOISE_KEY_2 = new Uint8Array(
+  ((): number[] => {
+    const b = [];
+    for (let i = 0; i < 32; i++) {
+      b.push(2);
+    }
+    return b;
+  })()
+);