+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Replaced `rlp` dependency with `@ethersproject/rlp`.
+- **breaking**: `staticNoiseKey` changed from `Buffer` to `Uint8Array`.
+
+### Removed
+
+- `base64url` and `bigint-buffer` dependencies.
+
 ## [0.17.0] - 2022-02-16
 
 ### Changed
@@ -21,7 +30,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `axios` dependency in favour of fetch.
-- `base64url` and `bigint-buffer` dependencies.
 
 ## [0.16.0] - 2022-01-31